Renamed hazards namespace to singular form for consistency.
