+++ Source/NexusForever.Network.World/Message/Model/Hazard/HazardModifier.cs
@@ -1,18 +1,18 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Hazards
+namespace NexusForever.Network.World.Message.Model.Hazard
 {
     public class HazardModifier : IWritable
     {
         public float Multiplier { get; set; }
         public float Offset { get; set; }
-        public bool bSuspended { get; set; }
+        public bool Suspended { get; set; }
 
         public void Write(GamePacketWriter writer)
         {
             writer.Write(Multiplier);
             writer.Write(Offset);
-            writer.Write(bSuspended);
+            writer.Write(Suspended);
         }
     }
 }
