+++ Source/NexusForever.Network.World/Message/Model/Hazard/ServerHazardList.cs
@@ -1,6 +1,6 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Hazards
+namespace NexusForever.Network.World.Message.Model.Hazard
 {
     [Message(GameMessageOpcode.ServerHazardList)]
     public class ServerHazardList : IWritable
