+++ Source/NexusForever.Network.World/Message/Model/Hazard/ServerHazardModifiers.cs
@@ -1,6 +1,6 @@
 ﻿using NexusForever.Network.Message;
 
-namespace NexusForever.Network.World.Message.Model.Hazards
+namespace NexusForever.Network.World.Message.Model.Hazard
 {
     [Message(GameMessageOpcode.ServerHazardModifiers)]
     public class ServerHazardModifiers : IWritable
